Initial croquet application
This commit adds a base application with a croquet
interface that allows you to register a-frame components
with croquet and reflect that to other instances of the application.
Also included is very basic garden generation to test
croquet and a basic "player" that you can move around
in and others can see.